Refactor and fix icon locations
This refactors out the duplicated desktop launchers, replacing them with
a template which is used by the Makefile to generate a launcher for each
application.

It also fixes the default icon installation location to remove the
implicit dependency on `hicolor-icon-theme` and place them in the
default icon search path.

Packagers could use e.g. imagemagick to rescale the 1024x1024 icon and
install smaller icons into the appropriate hicolor icon locations.