Refactor Makefile to remove all duplication
This is a fairly major refactor of the Makefile content to entirely
remove duplication of commands, as well as ensuring all variable
placeholders are populated or blanked appropriately (previously,
placeholders were not always replaced and so the raw text was left
in place).

It also corrects the use of a @PREFIX@ holder in the ms-office-online
script where /usr/bin/env is instead "always" correct.